TAI Network Interface Bit Error Rate attributes
This patch adds two attributes to the network interface object. The first is the
current bit error rate, or BER. This is most recently calculated BER and is
calculated for errors that occur prior to any error correction, such as FEC.
This is a floating point number.

The second attribute if the amount of time over which the BER data for the
previous attribute was collected. This is a 32 bit unsigned value in units of
microseconds. For example, if error data was collected over a period of 10
milliseconds, this attribute would contain 10000.

Signed-off-by: Scott Emery <scotte@cumulusnetworks.com>