TAI: Allocate custom attribute range for the AC400 adapter
This patch adds two values to each of the three objects: a start and end
of the custom attribute values which are reserved for the AC400 adapter.
As discussed in the TAI workshop in London on October 18, 2018, each adapter
implementation which implements custom attributes should reserve 64K attribute
values in the custom range. The adapter is then free to implement up to 64K
custom attributes, which should be far above the number required.

Signed-off-by: Scott Emery <scotte@cumulusnetworks.com>