framework: compare with default value
the previous behavior only do comparison when the value is set by a
user and the value was pushed to diff array even if it is equal to the
default value.

this commit fix the behavior by using Config::_get() which returns a
default value when the value is not set by a user.

Signed-off-by: Wataru Ishida <ishida@nel-america.com>