inc: add an attribute for implementation templates
Implementation templates represent the list of transmission modes
supported by the network interface. Implementation templates are
described as a list of enum lists, and each transmission mode is
expressed as a combination of enum values. Each enum list has an enum
id (attribute id) and a list of enum values. The i-th transmission
mode is a combination of i-th values of each enum list.

Implementation templates also represent the preference among the
transmission modes. When only a part of the parameters corresponds to
the transmission mode is configured, then the first mode that matches
the current configured value will be used. Therefore the head of the
list is the default mode.